Исправлено поведение IAsset при копировании и перемещении Добавлены методы GetViewersRegistry и GetAssetViewersRegistry в Editor Добавлено поле SelectedAsset в Editor Реализован AssetsEditor, AssetViewersRegister, IAssetViewer Частично реализован MaterialAssetViewer Изменено поведение MaterialViewer при вызове редактора материала В RenderSystem встроена проверка наличия шейдера

+++ include/Editor/UI/EditorViewer.h
@@ -61,7 +61,7 @@ public:
 
     EditorCamera EditorCamera;
     f32 CameraSpeed = 5.0f;
-    f32 CameraSense = 5.0f;
+    f64 CameraSense = 5.0f;
     bool CameraInverted = false;
 
     void Draw() final;
